venv_requirements.txt
tqdm==4.65.0
Pillow
imageio
matplotlib
kornia
natsort
pyyaml
wandb
lpips
torchmetrics
cyclonedds
pytorch-msssim
plyfile==0.8.1
opencv-python
plotly
# git+https://github.com/JonathonLuiten/diff-gaussian-rasterization-w-depth.git@cb65e4b86bc3bd8ed42174b72a62e8d3a3a71110